everest-core: add patch for System module
The downloads for Tarragon platform are quite large at the moment,
this we build developer images. The file might exceed the available
space on the /tmp ramfs, so it is better to download to /srv partition
instead.
This commit introduces a patch for this, a clean approach should
be developed with upstream.

Signed-off-by: Michael Heimpold <michael.heimpold@chargebyte.com>