everest-core: prefer journald timestamps
This adds a patch to everest-core which drop the timestamp from
the default logging configuration.

On chargebyte's platforms, we use journald for logging which adds
a high-resolution timestamp and other meta data to each line.

For developer convenience, we create an alias in the shell so
that the high-precision timestamp is displayed by default.

Signed-off-by: Michael Heimpold <michael.heimpold@chargebyte.com>