core-bundle: silence warning message when copying mosquitto configs
In case customer has not added mosquitto config sniplets, currently
the following warning can be seen in the rauc trace (wrapped for
readability):

-snap-
run-parts: /run/rauc/bundle/post-install.d/20-copy-mosquitto-user-configs
 exited with return code 2
-snap-

This is due to the used ls command which contributes to the overall exit
code of the whole script - which is not desired.
Using a full if-then construct prevents this potentially confusing message.

Signed-off-by: Michael Heimpold <michael.heimpold@chargebyte.com>